requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pymysql==1.1.0
requests==2.31.0
openai>=1.0.0
rapidfuzz>=3.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
